@@ -1,6 +1,6 @@
 module go.jetpack.io/devbox
 
-go 1.23.0
+go 1.24.0
 
 require (
 	github.com/AlecAivazis/survey/v2 v2.3.7
